Move eccvm_recursion
