--- barretenberg/cpp/src/barretenberg/eccvm_recursion/CMakeLists.txt
@@ -1,1 +1,0 @@
-barretenberg_module(eccvm_recursion eccvm stdlib_honk_verifier)
