+++ barretenberg/cpp/src/barretenberg/stdlib/CMakeLists.txt
@@ -2,6 +2,7 @@ add_subdirectory(hash)
 add_subdirectory(client_ivc_verifier)
 add_subdirectory(commitment)
 add_subdirectory(encryption)
+add_subdirectory(eccvm_verifier)
 add_subdirectory(goblin_verifier)
 add_subdirectory(primitives)
 add_subdirectory(honk_verifier)
