+++ barretenberg/cpp/src/barretenberg/stdlib/eccvm_verifier/CMakeLists.txt
@@ -1,0 +1,1 @@
+barretenberg_module(stdlib_eccvm_verifier eccvm stdlib_honk_verifier)
