+++ barretenberg/cpp/src/barretenberg/stdlib/goblin_verifier/CMakeLists.txt
@@ -1,8 +1,8 @@
 barretenberg_module(
     stdlib_goblin_verifier
-    stdlib_circuit_builders
-    translator_vm_recursion
-    eccvm_recursion
-    stdlib_honk_verifier
     goblin
+    stdlib_circuit_builders
+    stdlib_honk_verifier
+    stdlib_eccvm_verifier
+    translator_vm_recursion
 )
