+++ barretenberg/cpp/src/barretenberg/stdlib/protogalaxy_verifier/CMakeLists.txt
@@ -1,6 +1,6 @@
 barretenberg_module(
     stdlib_protogalaxy_verifier
-    stdlib_honk_verifier 
-    stdlib_transcript
+    stdlib_goblin_verifier 
+    stdlib_blake3s
     protogalaxy
 )
